Move Spree::BackendConfiguration::MenuItem to its own file
